Start splitting TestCase for plugins and libraries
The tests specific for plugin scanning are very useful, but they need to
be run in a different way compared to the per-plugin instance tests.
Testing the scanning performance for an entire plugin library again and
again for every plugin in the library doesn't make any sense, and that
tests requires the library to not be initialized yet. Hence the need for
two different sets of tests.